Removed a lot of the repitition.
In the header of the products page a link is created per product.
Caffeine is set as the home and active link. Each link is connected
to a tab panel that contains the same product title ID. The ID uses the products title
and if there is a space in the title it is cut. Need to look into
pre setting the order of the products displayed.

In the products.js file every product now uses the same add,remove, and select listener.
Each listener modifies the appropriate produce price being displayed based upon the ID of the
variation and the products title.

on the main product page( energy.html, need to change the name of this) a tab panel
is created for each product other than caffeine because caffeine is the default tab that
is active when the page is inititally opened. The ID per each tab is the products title and
similar to the link that is connected to each products tab panel the ID is the products title.
If the product title has a space it is stripped from the title.